--- Multi-agent_RL/actor.py
@@ -1,60 +1,0 @@
-import os.path
-import numpy as np
-from tensorflow.keras import layers, models, optimizers
-from tensorflow.keras import backend as K
-
-
-class Actor:
-    """Actor (Policy) Model."""
-
-    def __init__(self, state_size, action_size):#, action_low, action_high):
-        """Initialize parameters and build model.
-
-        Params
-        ======
-            state_size (int): Dimension of each state
-            action_size (int): Dimension of each action
-            action_low (array): Min value of each action dimension
-            action_high (array): Max value of each action dimension
-        """
-        self.state_size = state_size
-        self.action_size = action_size
-        #self.action_low = action_low
-        #self.action_high = action_high
-        #self.action_range = self.action_high - self.action_low
-
-        self.build_model()
-
-    def build_model(self):
-        """Build an actor (policy) network that maps states -> actions."""
-        # Define input layer (states)
-        states = layers.Input(shape=(self.state_size,), name='states')
-
-        # Add hidden layers
-        net = layers.Dense(units=512, activation='relu')(states)
-        net = layers.Dense(units=512, activation='relu')(net)
-        net = layers.Dense(units=512, activation='relu')(net)
-
-        # Add final output layer with sigmoid activation
-        action1 = layers.Dense(units=1, activation='sigmoid')(net)
-        action2 = layers.Dense(units=1, activation='tanh')(net)
-
-        actions = layers.Concatenate()([action1, action2])
-
-        #raw_actions = layers.Dense(units=self.action_size, activation='sigmoid', name='raw_actions')(net)
-
-        # Scale [0, 1] output for each action dimension to proper range
-        #actions = layers.Lambda(lambda x: (x * self.action_range) + self.action_low, name='actions')(raw_actions)
-
-        # Create Keras model
-        self.model = models.Model(inputs=states, outputs=actions)
-
-        # Define loss function using action value (Q value) gradients
-        action_gradients = layers.Input(shape=(self.action_size,))
-        loss = K.mean(-action_gradients * actions)
-
-        # Define optimizer and training function
-        optimizer = optimizers.Adam(lr=0.001)
-        updates_op = optimizer.get_updates(params=self.model.trainable_weights, loss=loss)
-        self.train_fn = K.function(inputs=[self.model.input, action_gradients, K.learning_phase()],
-                                   outputs=[loss], updates=updates_op)
